Thunkified generateSchedules
